backend-master/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
click==7.1.2
Django==3.1.5
django-cors-headers==3.7.0
django-elasticsearch-dsl==7.1.4
django-elasticsearch-dsl-drf==0.22
django-nine==0.2.4
djangorestframework==3.12.2
elasticsearch==7.12.0
elasticsearch-dsl==7.3.0
jsonpickle==2.0.0
Markdown==3.0
psycopg2==2.8.6
PyJWT==1.0.0
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
stock==0.1
urllib3==1.26.4
pandas==1.1.5
numpy==1.19.5
scikit-learn==0.24.2
nltk==3.5.00
